EC-10 fixed three JS files for three products page
