+++ children.html
@@ -46,5 +46,6 @@
     -->
         </section>
     </main>
+    <script type="module" src="js/children.js"></script>
 </body>
 </html>
