+++ js/men.js
@@ -1,0 +1,2 @@
+import { getProducts, productList, displayProducts, filterAndDisplay} from "./fetchProductData.js";
+filterAndDisplay("men");
