+++ men.html
@@ -43,6 +43,7 @@
           </div>
     -->
         </section>
+        <script type="module" src="js/men.js"></script>
     </main>
 </body>
 </html>
